+++ crates/ghost_actor_doc_example/src/lib.rs
@@ -9,33 +9,6 @@ pub enum MyError {
 }
 
 crate::ghost_chan! {
-    /// Test multi-line doc on channel type.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// // we can even run doc-tests
-    /// assert_eq!(true, true);
-    /// ```
-    pub chan MyCustomChan<MyError> {
-        /// this function has no inputs or outputs -- not all that useful
-        fn no_input_no_output() -> ();
-
-        /// this function has one input and no outputs
-        fn one_input_no_outpun(i: u32) -> ();
-
-        /// this function has no inputs and one output
-        fn no_input_one_output() -> u32;
-
-        /// 1 and 1
-        fn one_input_one_output(i: u32) -> u32;
-
-        /// 2 and 2
-        fn two_inputs_two_outputs(i: u32, j: u32) -> (u32, u32);
-    }
-}
-
-crate::ghost_actor! {
     /// Test multi-line doc on actor type.
     ///
     /// # Example
@@ -44,7 +17,7 @@ crate::ghost_actor! {
     /// // we can even run doc-tests
     /// assert_eq!(true, true);
     /// ```
-    pub actor MyActor<MyError> {
+    pub chan MyActor<MyError> {
         /// this function has no inputs or outputs -- not all that useful
         fn no_input_no_output() -> ();